move hook details to separate component/file
